Add activity CRUD + input valdiation

+++ app/Http/Controllers/DailyActivitiesController.php
@@ -8,6 +8,16 @@ use App\DailyActivity;
 class DailyActivitiesController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -26,7 +36,7 @@ class DailyActivitiesController extends Controller
      */
     public function create()
     {
-        //
+        // return create view
         return view('daily-activities.create');
     }
 
@@ -38,24 +48,31 @@ class DailyActivitiesController extends Controller
      */
     public function store(Request $request)
     {
-        //
+        // Validate the input
         $this->validate($request,[
             'physical_activity' => 'required',
+            'hours' => 'nullable|integer|between:0,24',
+            'minutes' => 'nullable|integer|between:0,60',
             'fruit_vege' => 'required',
-            'smoke' => 'required'
+            'servings' => 'nullable|integer|between:0,99',
+            'smoke' => 'required',
+            'date' => 'required|unique:daily_activities,date,NULL,id,user_id,'.\Auth::id()
         ]);
         
+        // Create new object and assign their values from form input
         $daily_activity = new DailyActivity;
+        $daily_activity->user_id = auth()->user()->id;
+        $daily_activity->date = $request->input('date');
         $daily_activity->physical_activity = $request->input('physical_activity');
         $daily_activity->hours = $request->input('hours');
         $daily_activity->minutes = $request->input('minutes');
         $daily_activity->fruit_vege = $request->input('fruit_vege');
         $daily_activity->servings = $request->input('servings');
         $daily_activity->smoke = $request->input('smoke');
-        $daily_activity->user_id = 1;
+        // Save the record in database
         $daily_activity->save();
 
-        return redirect('daily-activities')->with('success', 'Activity added');
+        return redirect('home')->with('success', 'Activity added');
 
     }
     
@@ -68,7 +85,7 @@ class DailyActivitiesController extends Controller
      */
     public function show($id)
     {
-        //
+        // return show view
         $daily_activity =  DailyActivity::find($id);
         return view('daily-activities.show')->with('daily_activity', $daily_activity);
     }
@@ -81,8 +98,12 @@ class DailyActivitiesController extends Controller
      */
     public function edit($id)
     {
-        //
+        // find activity based on id
         $daily_activity =  DailyActivity::find($id);
+        // Check for correct user
+        if(auth()->user()->id !== $daily_activity->user_id){
+            return redirect('home')->with('error', 'Unauthorized Page');
+        }
         return view('daily-activities.edit')->with('daily_activity', $daily_activity);
     }
 
@@ -95,7 +116,29 @@ class DailyActivitiesController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
+        // validate input
+        $this->validate($request,[
+            'physical_activity' => 'required',
+            'hours' => 'nullable|integer|between:0,24',
+            'minutes' => 'nullable|integer|between:0,60',
+            'fruit_vege' => 'required',
+            'servings' => 'nullable|integer|between:0,99',
+            'smoke' => 'required',
+            'date' => 'required|unique:daily_activities,date,NULL,id,user_id,'.\Auth::id()
+        ]);
+        
+        // Update the record based on the id
+        $daily_activity = DailyActivity::find($id);
+        $daily_activity->physical_activity = $request->input('physical_activity');
+        $daily_activity->hours = $request->input('hours');
+        $daily_activity->minutes = $request->input('minutes');
+        $daily_activity->fruit_vege = $request->input('fruit_vege');
+        $daily_activity->servings = $request->input('servings');
+        $daily_activity->smoke = $request->input('smoke');
+        // Save the record in database
+        $daily_activity->save();
+
+        return redirect('home')->with('success', 'Activity updated');
     }
 
     /**
@@ -106,9 +149,14 @@ class DailyActivitiesController extends Controller
      */
     public function destroy($id)
     {
-        //
+        // find record based on id
         $daily_activity = DailyActivity::find($id);
+        // Check for correct user
+        if(auth()->user()->id !== $daily_activity->user_id){
+            return redirect('home')->with('error', 'Unauthorized Page');
+        }
+        // delete the record in database
         $daily_activity->delete();
-        return redirect('daily-activities')->with('success', 'Activity removed');
+        return redirect('home')->with('success', 'Activity removed');
     }
 }
